fix: Split tests in 2 files
For better readability, the tests with and
without recurrence are separated
